// src/app/shared/enums/notification-type.enum.ts
export enum NotificationTypeEnum {
    // Navigation Types
    RESERVATION = 'reservation',
    PODCAST = 'podcast',
    COMMUNITY = 'community',
    TASK = 'task',
    OFFER = 'offer',
    SESSION_REMINDER = 'session_reminder',
    NEW_MESSAGE = 'new_message',
    APPOINTMENT = 'appointment',

    // Status Types
    SUCCESS = 'success',
    WARNING = 'warning',
    ERROR = 'error',
    INFO = 'info',
    GENERAL = 'general'
}